Initial commit
The main branch contains update-to-head.sh, used to sync from upstream.

This script is a copy of others like it in the mid/downstream knative
repositories in the openshift organization. It should be run daily as
an automated task. At the moment, however, we don't have that task set
up.

The script's purpose is to mirror the upstream source repository - in
this case, knative-sandbox/kn-plugin-func - to the `release-next` branch
and then trigger a CI run on the `release-next-ci` branch, by adding a
datestamped `ci` file.

Signed-off-by: Lance Ball <lball@redhat.com>